+++ packages/quizzy/.editorconfig
@@ -1,0 +1,16 @@
+root = true
+
+#best practise is lf, sorry notepad users
+[*]
+end_of_line = lf
+insert_final_newline = true
+indent_style = space
+indent_size = 4
+quote_type = single
+trim_trailing_whitespace = true
+
+[*.{css,scss,less,sass}]
+indent_style = space
+indent_size = 2
+quote_type = single
+trim_trailing_whitespace = true
