+++ packages/quizzy/package.json
@@ -1,28 +1,17 @@
 {
-    "name": "quizzy",
+    "name": "@sbb-mono/quizzy",
     "version": "0.1.0",
     "private": true,
     "license": "MIT",
     "dependencies": {
-        "@testing-library/jest-dom": "^5.11.4",
-        "@testing-library/react": "^10.4.9",
-        "@testing-library/user-event": "^7.1.2",
-        "@types/jest": "^24.0.0",
-        "@types/node": "^12.0.0",
-        "@types/react": "^16.9.0",
-        "@types/react-dom": "^16.9.0",
+        "@sbb-mono/ui": "1.0.0",
         "react": "^16.13.1",
         "react-app-polyfill": "^1.0.6",
         "react-dom": "^16.13.1",
-        "react-scripts": "3.4.3",
-        "typescript": "^4.0.2",
-        "webpack": "4.42.0"
+        "react-scripts": "3.4.3"
     },
     "scripts": {
-        "start": "react-scripts start",
-        "build": "react-scripts build",
-        "test": "react-scripts test",
-        "eject": "react-scripts eject",
+        "start": "parcel src/index.html --open",
         "lint": "lint-staged",
         "lint-script": "eslint \"src/**/*.{js,jsx,ts,tsx}\" --fix",
         "lint-style": "stylelint \"src/**/*.scss\" --fix"
@@ -44,8 +33,20 @@
         ]
     },
     "devDependencies": {
+        "@babel/core": "^7.11.6",
+        "@babel/preset-env": "^7.11.5",
+        "@babel/preset-react": "^7.10.4",
+        "@testing-library/jest-dom": "^5.11.4",
+        "@testing-library/react": "^10.4.9",
+        "@testing-library/user-event": "^7.1.2",
+        "@types/jest": "^24.0.0",
+        "@types/node": "^12.0.0",
+        "@types/react": "^16.9.0",
+        "@types/react-dom": "^16.9.0",
         "@typescript-eslint/eslint-plugin": "^3.10.1",
         "@typescript-eslint/parser": "^3.10.1",
+        "babel-preset-env": "^1.7.0",
+        "babel-preset-react": "^6.24.1",
         "eslint-config-airbnb": "18.2.0",
         "eslint-config-airbnb-typescript": "^9.0.0",
         "eslint-config-prettier": "^6.11.0",
@@ -58,12 +59,14 @@
         "husky": "^4.2.5",
         "lint-staged": "^10.2.13",
         "mini-css-extract-plugin": "^0.10.0",
+        "parcel-bundler": "^1.12.4",
         "postcss-loader": "^3.0.0",
         "sass-loader": "^9.0.3",
         "style-loader": "^1.2.1",
         "stylelint": "^13.7.0",
         "stylelint-config-sass-guidelines": "^7.1.0",
-        "stylelint-scss": "^3.18.0"
+        "stylelint-scss": "^3.18.0",
+        "typescript": "^4.0.2"
     },
     "husky": {
         "hooks": {
