+++ packages/quizzy/src/index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <title>parcel-typescript-template</title>
+  </head>
+  <body>
+    <noscript>You need to enable Javascript to run this application.</noscript>
+    <div id="quiz" data-api="http://my-example-api.com"></div>
+    <script src="../src/index.tsx"></script>
+  </body>
+</html>
