+++ packages/quizzy/src/quiz.json
@@ -1,6 +1,6 @@
 {
     "quiz": {
-        "title": "Selbsttest",
+        "title": "Selbsttest Lia",
         "questions": [
             {
                 "id": 100,
